Restructured the api
