+++ README.md
@@ -35,13 +35,22 @@ The project is designed as one deployable app: Vite builds the frontend into `di
 ```text
 .
 ├── api/
-│   ├── server.js              # Express server and API routes
+│   ├── server.js              # Compatibility wrapper for api/src/server.js
+│   ├── src/
+│   │   ├── app.js             # Express app, middleware, static hosting, route registration
+│   │   ├── server.js          # HTTP server startup
+│   │   ├── config/            # Environment and path configuration
+│   │   ├── controllers/       # HTTP request/response handlers
+│   │   ├── middleware/        # Logging, uploads, and error handling
+│   │   ├── routes/            # Endpoint definitions
+│   │   ├── services/          # Provider integrations and feature logic
+│   │   └── utils/             # Shared backend utilities
 │   ├── package.json           # API-specific dependencies/scripts
-│   ├── chatGPT.js             # AI response helpers
-│   ├── gemini.js              # Gemini model and prompt helpers
-│   ├── youtube.js             # YouTube API helpers
-│   ├── transcriptService.js   # YouTube transcript helper
-│   ├── supadata.js            # Supadata transcript helper
+│   ├── chatGPT.js             # Legacy re-export for service module
+│   ├── gemini.js              # Legacy re-export for service module
+│   ├── youtube.js             # Legacy re-export for service module
+│   ├── transcriptService.js   # Legacy re-export for service module
+│   ├── supadata.js            # Legacy re-export for service module
 │   └── logs/                  # Runtime API logs
 ├── shared/
 │   └── cpuSimulator.js        # CPU simulator/compiler logic
@@ -74,13 +83,13 @@ PORT=8080
 GEMINI_API_KEY=your_gemini_api_key
 
 # OpenAI, if used by helper modules
-OPENAI_API_KEY=your_openai_api_key
+CHAT_GPT_API_KEY=your_openai_api_key
 
 # YouTube API
 YOUTUBE_API_KEY=your_youtube_api_key
 
 # Supadata
-SUPADATA_API_KEY=your_supadata_api_key
+SUPA_DATA_API_KEY=your_supadata_api_key
 
 # Google Cloud Text-to-Speech
 GOOGLE_APPLICATION_CREDENTIALS=/path/to/service-account.json
@@ -360,7 +369,7 @@ Root `package.json`:
 ```bash
 npm run dev      # Start Vite dev server
 npm run build    # Build frontend
-npm run start    # Start Express server from api/server.js
+npm run start    # Start Express server from api/src/server.js
 npm run preview  # Preview built Vite app
 npm run lint     # Run ESLint
 ```
@@ -368,9 +377,19 @@ npm run lint     # Run ESLint
 API `package.json`:
 
 ```bash
-npm start        # Start api/server.js
+npm start        # Start api/src/server.js
 npm run debug    # Start API with Node inspector and watch mode
 ```
+
+## Adding Backend Features
+
+Add new backend behavior in the same dependency direction as the existing modules:
+
+1. Add the URL and HTTP method in `api/src/routes/<feature>.routes.js`.
+2. Put request parsing and response shaping in `api/src/controllers/<feature>.controller.js`.
+3. Put external API calls or feature logic in `api/src/services/<feature>.service.js`.
+4. Register the route module in `api/src/routes/index.js`.
+5. Add new configuration names to `api/src/config/env.js` and document blank values in `api/.env-template`.
 
 ## Deployment
