+++ api/.env-template
@@ -1,3 +1,10 @@
 GEMINI_API_KEY=
+GEMINI_MODEL=
 YOUTUBE_API_KEY=
 SUPA_DATA_API_KEY=
+CHAT_GPT_API_KEY=
+GOOGLE_APPLICATION_CREDENTIALS=
+GOOGLE_APPLICATION_CREDENTIALS_JSON=
+TTS_SA_JSON=
+PDF_UPLOAD_MAX_MB=
+PORT=
