+++ api/src/config/paths.js
@@ -1,0 +1,12 @@
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+export const apiSrcDir = path.dirname(fileURLToPath(import.meta.url));
+export const apiRoot = path.resolve(apiSrcDir, '../..');
+export const projectRoot = path.resolve(apiRoot, '..');
+export const distDir = path.join(projectRoot, 'dist');
+export const uploadsDir = path.join(projectRoot, 'var/uploads');
+export const secretsDir = path.join(projectRoot, 'var/secrets');
+export const logsDir = path.join(apiRoot, 'logs');
+export const errorLogPath = path.join(apiRoot, 'error.log');
+export const pythonYoutubeServicePath = path.join(apiRoot, 'youtube_service.py');
