+++ api/src/constants/routes.js
@@ -1,0 +1,1 @@
+export const API_PREFIX = '/api';
