+++ api/src/server.js
@@ -1,0 +1,6 @@
+import app from './app.js';
+import { env } from './config/env.js';
+
+app.listen(env.port, () => {
+  console.log(`Server listening on port ${env.port}`);
+});
