+++ api/src/utils/asyncHandler.js
@@ -1,0 +1,3 @@
+export const asyncHandler = (handler) => (req, res, next) => {
+  Promise.resolve(handler(req, res, next)).catch(next);
+};
