+++ api/src/utils/httpErrors.js
@@ -1,0 +1,8 @@
+export class HttpError extends Error {
+  constructor(statusCode, message) {
+    super(message);
+    this.statusCode = statusCode;
+  }
+}
+
+export const badRequest = (message) => new HttpError(400, message);
