+++ package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "start": "node api/server.js",
+    "start": "node api/src/server.js",
     "preview": "vite preview"
   },
   "proxy": "http://localhost:3005",
